refactor(bazel): cleanup
