+++ detectable/src/test/java/com/synopsys/integration/detectable/detectables/bazel/functional/bazel/pipeline/step/HaskellCabalLibraryJsonProtoParserTest.java
@@ -11,7 +11,7 @@ import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.Test;
 
 import com.google.gson.Gson;
-import com.synopsys.integration.detectable.detectables.bazel.pipeline.step.HaskellCabalLibraryJsonProtoParserSimple;
+import com.synopsys.integration.detectable.detectables.bazel.pipeline.step.HaskellCabalLibraryJsonProtoParser;
 import com.synopsys.integration.exception.IntegrationException;
 import com.synopsys.integration.util.NameVersion;
 
@@ -22,7 +22,7 @@ public class HaskellCabalLibraryJsonProtoParserTest {
         File jsonProtoFile = new File("src/test/resources/detectables/functional/bazel/jsonProtoForHaskellCabalLibraries.txt");
         String jsonProtoHaskellCabalLibrary = FileUtils.readFileToString(jsonProtoFile, StandardCharsets.UTF_8);
 
-        HaskellCabalLibraryJsonProtoParserSimple parser = new HaskellCabalLibraryJsonProtoParserSimple(new Gson());
+        HaskellCabalLibraryJsonProtoParser parser = new HaskellCabalLibraryJsonProtoParser(new Gson());
 
         List<NameVersion> dependencies = parser.parse(jsonProtoHaskellCabalLibrary);
         assertEquals(1, dependencies.size());
